Menambahkan halaman edit user dan add user

+++ app/Http/Controllers/BarangController.php
@@ -3,57 +3,67 @@
 namespace App\Http\Controllers;
 
 use App\Models\Barang;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
 
 class BarangController extends Controller
 {
-    public function index()
-    {
-        $stocks = Barang::select('id', 'nama', 'deskripsi', 'stock')->get();
-
-        return view('home', compact('stocks'));
-    }
 
     public function create()
     {
-        return view('barang.create');
+        if (Auth::check() && trim(Auth::user()->role) == 'admin') {
+
+            return view('barang.create');
+        } else {
+            return view('home');
+        }
     }
 
     public function destroy($id)
     {
-        $barang = Barang::findOrFail($id);
+        if (Auth::check() && trim(Auth::user()->role) == 'admin') {
 
-        // Hapus file gambar dari storage
-        if (Storage::disk('public')->exists($barang->img_path)) {
-            Storage::disk('public')->delete($barang->img_path);
+            $barang = Barang::findOrFail($id);
+
+            // Hapus file gambar dari storage
+            if (Storage::disk('public')->exists($barang->img_path)) {
+                Storage::disk('public')->delete($barang->img_path);
+            }
+
+            // Hapus data barang dari database
+            $barang->delete();
+
+            return redirect()->route('barang')->with('success', 'Barang dan gambar berhasil dihapus.');
+        } else {
+            return view('home');
         }
-
-        // Hapus data barang dari database
-        $barang->delete();
-
-        return redirect()->route('barang')->with('success', 'Barang dan gambar berhasil dihapus.');
     }
 
     public function store(Request $request)
     {
-        $request->validate([
-            'nama' => 'required|max:50',
-            'deskripsi' => 'required',
-            'stock' => 'required|integer|min:0',
-            'gambar' => 'required|image|mimes:jpg,jpeg,png,gif|max:2048',
-        ]);
+        if (Auth::check() && trim(Auth::user()->role) == 'admin') {
 
-        // Simpan gambar ke folder storage/app/public/barang/
-        $path = $request->file('gambar')->store('barang', 'public');
+            $request->validate([
+                'nama' => 'required|max:50',
+                'deskripsi' => 'required',
+                'stock' => 'required|integer|min:0',
+                'gambar' => 'required|image|mimes:jpg,jpeg,png,gif|max:2048',
+            ]);
 
-        Barang::create([
-            'nama' => $request->nama,
-            'deskripsi' => $request->deskripsi,
-            'stock' => $request->stock,
-            'img_path' => $path,
-        ]);
+            // Simpan gambar ke folder storage/app/public/barang/
+            $path = $request->file('gambar')->store('barang', 'public');
 
-        return redirect()->back()->with('success', 'Barang berhasil ditambahkan!');
+            Barang::create([
+                'nama' => $request->nama,
+                'deskripsi' => $request->deskripsi,
+                'stock' => $request->stock,
+                'img_path' => $path,
+            ]);
+
+            return redirect()->back()->with('success', 'Barang berhasil ditambahkan!');
+        } else {
+            return view('home');
+        }
     }
 }
